refactor header includes; update control socket logic for advanced commands
