+++ src/iad/audio/output.h
@@ -1,9 +1,7 @@
 #ifndef OUTPUT_H
 #define OUTPUT_H
 
-#include <imp/imp_audio.h>
-#include <imp/imp_log.h>
-#include <../include/cJSON.h>
+#include <imp/imp_audio.h>  // for AUDIO_SAMPLE_RATE_48000
 
 #define DEFAULT_AO_SAMPLE_RATE AUDIO_SAMPLE_RATE_48000
 #define DEFAULT_AO_MAX_FRAME_SIZE 1280
@@ -14,46 +12,13 @@
 #define DEFAULT_AO_DEV_ID 0
 #define DEFAULT_AO_CHN_ID 0
 
-typedef struct {
-    cJSON *samplerateItem;
-    cJSON *bitwidthItem;
-    cJSON *soundmodeItem;
-    cJSON *frmNumItem;
-    cJSON *chnCntItem;
-    cJSON *SetVolItem;
-    cJSON *SetGainItem;
-} AudioOutputAttributes;
-
-typedef struct {
-    cJSON *aoDevIDItem;
-    cJSON *channel_idItem;
-} PlayAttributes;
-
 // Functions
 void reinitialize_audio_device(int aoDevID, int aoChnID);
 void *ao_test_play_thread(void *arg);
-void pause_audio_output(void);
-void clear_audio_output_buffer(void);
-void resume_audio_output(void);
-void flush_audio_output_buffer(void);
-void mute_audio_output_device(int mute_enable);
-
-AudioOutputAttributes get_audio_attributes(void);
-void free_audio_attributes(AudioOutputAttributes *attrs);
-PlayAttributes get_audio_play_attributes(void);
-void free_audio_play_attributes(PlayAttributes *attrs);
 
 extern int g_ao_max_frame_size;
 void set_ao_max_frame_size(int frame_size);
 void cleanup_audio_output();
 int disable_audio_output(void);
-
-
-/**
- * Retrieves audio attributes (device and channel IDs) either from PlayAttributes or defaults.
- * @param aoDevID Pointer to store the retrieved Device ID.
- * @param aoChnID Pointer to store the retrieved Channel ID.
- */
-void get_audio_device_attributes(int *aoDevID, int *aoChnID);
 
 #endif // OUTPUT_H
