+++ src/iad/network/network.h
@@ -1,10 +1,6 @@
 #ifndef NETWORK_H
 #define NETWORK_H
 
-#include <sys/socket.h>
-#include <sys/un.h>
-#include <unistd.h>
-
 extern char AUDIO_INPUT_SOCKET_PATH[];
 extern char AUDIO_OUTPUT_SOCKET_PATH[];
 extern char AUDIO_CONTROL_SOCKET_PATH[];
@@ -14,6 +10,15 @@ extern char AUDIO_CONTROL_SOCKET_PATH[];
 
 #define AUDIO_INPUT_REQUEST 1
 #define AUDIO_OUTPUT_REQUEST 2
+
+// Control commands
+#define CONTROL_GET_COMMAND 3
+#define CONTROL_SET_COMMAND 4
+
+// Response codes
+#define RESPONSE_OK 200
+#define RESPONSE_ERROR 400
+#define RESPONSE_UNKNOWN_VARIABLE 404
 
 // Functions
 void *audio_input_server_thread(void *arg);
