+++ src/iad/utils/config.h
@@ -1,5 +1,6 @@
 #ifndef CONFIG_H
 #define CONFIG_H
+
 #include <cJSON.h>
 
 // -----------------------------
